Remove backup file on overwritten "File Save"
When you try to overwritten a file, dia create an copy of the original
file with '~' at the end. Some user, complained about this and Debian
package removed this feature. This commit was orignated from Debian
package follows the original message:

Dia normally saves a copy of an old saved file if it is overwritten via
the "File Save" function. This clutters the directory. Since the common
behaviour of programs is to not save backups in a separate location
(also for security reasons), this patch removes this feature and
therefore removes backup files when dia is sure to have the newly
written file in the correct location.

Bug-Debian: http://bugs.debian.org/173650

Signed-off-by: Roland Stigge <stigge@antcom.de>