detect freetype via pkg-config
This allows Dia to be built with newer Freetype libraries. These do
not include the `freetype-config` tool any more, which Dia's build
process used to detect installed Freetype headers and libraries.

With this patch Dia builds in Debian jessie as well as in Debian buster.

The patch comes from Lars Wendler <polynomial-c@gentoo.org> and was
applied to Debian by Gregor Herrmann in 0.97.3+git20160930-8.1.

https://bugs.debian.org/887602
https://gitlab.gnome.org/GNOME/dia/merge_requests/1
https://bugs.gentoo.org/654814